// src/app/shared/models/definition-chart-config.ts
export class DefinitionChartConfig {
    edit: boolean;
    showUserListProperty: boolean;
    showPredefinedListProperty: boolean;
    clickableUserListProperty: boolean;
    clickablePredefinedListProperty: boolean;
    showNonListProperty: boolean;
    clickableNonListProperty: boolean;
    showListMembers: boolean;
    clickableListMembers: boolean;
}